requirements.txt
fredapi>=0.5.1
pandas>=2.0
matplotlib>=3.7
python-dotenv>=1.0
PyYAML>=6.0
